Fix race and hang when fd gets closed
As the FIXME points out, the finalize function may be invoked from a
finalizer-thread. To avoid races with the scheduler thread, spawn a
fiber on the scheduler that handles the fd to cleanup the fd-waiters.

Also ensure that all waiters waiting on the closed fd get scheduled
again, as they would otherwise wait forever.